+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ## v2.15.0-dev.2
 ### Added
 - Added new package - [FramePack](https://github.com/lllyasviel/FramePack)
+- Added new package - [FramePack Studio](https://github.com/colinurbs/FramePack-Studio)
 ### Changed
 - Civitai model browser image loading now uses dynamic resizing for better performance and a smoother scrolling experience.
 - Detailed notifications for Civitai model browser api errors.